Move data classed to new DepartureDetails file
